backend/requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.1.4
django-cors-headers==4.6.0
django-debug-toolbar==5.0.1
djangorestframework==3.15.2
drf-spectacular==0.28.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
pillow==11.1.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.4.0
